+++ HvZ-TheGame/HvZ-TheGame/BaseClass.h
@@ -10,9 +10,9 @@ protected:
 public:
 	int itemGroup;
 	sf::Vector3i gridLocation;
-	sf::Vector3i pointLocation;
+	sf::Vector3f pointLocation;
 
-	BaseClass(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, sf::Vector3i pointLocation, int itemGroup);
+	BaseClass(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, sf::Vector3f pointLocation, int itemGroup);
 	~BaseClass();
 
 	virtual void move(int x, int y, int z);
