+++ HvZ-TheGame/HvZ-TheGame/Block.cpp
@@ -3,7 +3,7 @@
 #include "Constants.h"
 
 Block::Block(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, int blockType)
-	:BaseClass(world, texture, gridLocation, sf::Vector3i(0, 0, 0), BLOCK), blockType(blockType) {
+	:BaseClass(world, texture, gridLocation, sf::Vector3f(0, 0, 0), BLOCK), blockType(blockType) {
 	sprite.setOrigin(sprite.getLocalBounds().width / 2, sprite.getLocalBounds().height);
 	sprite.scale((BLOCK_SIZE * 2) / sprite.getLocalBounds().width, (BLOCK_SIZE * 2) / sprite.getLocalBounds().height);
 	sf::Vector3i p = cartesianToIsometric(gridLocation.x * BLOCK_SIZE, gridLocation.y * BLOCK_SIZE, gridLocation.z * BLOCK_SIZE);
